refactor(ers)!: remove single-creator Patreon support; claims-passthrough only (#24)
* refactor(ers)!: remove single-creator Patreon support; claims-passthrough only

Per product direction: the multi-creator SaaS path (claims-passthrough,
PR #23) is the only model, so the single-creator live-resolution path —
which required a per-deployment creator_access_token and a fixed
campaign_ids list — is removed entirely.

Deleted:
- client.go (708 lines): the Patreon v2 API client — member search by
  user id/email, /identity, client_credentials token refresh, JSON:API
  pagination/parsing.
- Config: api_base, token_url, client_id, client_secret,
  creator_access_token, campaign_ids.
- MembershipAPI interface + the live client field; NewClient/ClientOptions.
- resolveEntity/resolveFromClaims live-lookup branches (username/email/
  client-id/token); ErrPatreonUnavailable.

Kept (moved to types.go): Membership flattened view, status/tier
constants, normalizeStatus, slugify. resolveEntity now resolves only a
claims entity carrying arkavo_patreon; everything else is not-found (→
free when infer_unknown_as_free). The flattened .patreon.* view is still
surfaced for coarse legacy selectors.

Policy: removed the global `tier` hierarchy attribute + its four
.patreon.tier_slug subject mappings, and the hardcoded
campaign/value/arkavo mapping — both were the single-creator model.
campaign and campaign-tier are now dynamic (values: []), filled by the
ERS as direct entitlements per active membership. Onboarding a creator
needs zero policy changes.

Config examples + README rewritten for claims-passthrough; prod example
sets trust_materialized_claims + allow_direct_entitlements and drops the
creator credential / campaign placeholders.

BREAKING for any deployment relying on creator_access_token member
search — that mode no longer exists. Live-verified: passthrough decisions
correct with no Patreon credentials configured. Tests cover claims-only
resolution; go vet + golangci-lint clean.


* test: update attributes-endpoint test for removed tier attribute

The single-creator removal dropped the global `tier` attribute from
policy.patreon.yaml, but the attributes-endpoint test (PR #22) still
asserted tier was served. Go doesn't track the YAML as a test
dependency, so a stale local cache masked it; fresh CI caught it.

Switched the assertions to surviving creator-agnostic attributes:
classification (HIERARCHY) for the dereference + rule check, campaign
for the listing.


* docs(ers): update stale method comments for claims-passthrough-only model

ResolveEntities/CreateEntityChainsFromTokens doc comments still described
the removed live-Patreon lookup ("looks each entity up in Patreon",
"resolved via the JWT's identity hints"). Updated to reflect the
claims-only resolution, consistent with the resolveEntity/resolveFromClaims
docs updated in this PR. Review suggestion on #24, comment-only.


---------
