+++ service/entityresolution/patreon/v2/README.md
@@ -1,34 +1,33 @@
 # Patreon Entity Resolution Provider (v2)
 
-A Patreon-backed Entity Resolution Service (ERS) that resolves a subject's
-Patreon membership state and surfaces it as policy claims under the
-`.patreon.*` selector tree. Pairs with
+A Patreon-backed Entity Resolution Service (ERS) for the multi-creator SaaS
+model. It makes **no Patreon API calls**: identity.arkavo.net materializes
+every consumer's memberships into the `arkavo_patreon` CWT claim at token
+mint (using the consumer's own OAuth token, which sees every campaign they
+back), and this provider translates that trusted claim into entitlements.
+Pairs with
 [`examples/config/policy.patreon.yaml`](../../../../examples/config/policy.patreon.yaml).
 
 ## What it does
 
-For each `ResolveEntities` request entry, the provider looks the subject up in
-Patreon (by user id, email, JWT claim, or forwarded user OAuth token) and
-returns a representation that includes:
+When a subject's claims carry the materialized `arkavo_patreon` claim, the
+provider emits **campaign-qualified direct entitlements** per ACTIVE
+membership — in the creator's own tier vocabulary, with no per-creator
+config:
 
-```jsonc
-{
-  "patreon": {
-    "user_id": "12345",
-    "email": "fan@example.com",
-    "full_name": "Fan One",
-    "status": "active",            // active | declined | former
-    "tier_slug": "patron",         // slugified tier name; "free" if none
-    "tier_amount_cents": 500,
-    "campaign_ids": ["arkavo"],
-    "benefits": ["early-access", "exclusive-content"]
-  }
-}
+```
+https://patreon.arkavo.com/attr/campaign/value/<campaign_id>
+https://patreon.arkavo.com/attr/campaign-tier/value/<campaign_id>_<tier_slug>
 ```
 
-Subject mappings in `policy.patreon.yaml` key off these fields - e.g.
-`.patreon.tier_slug == "patron"` grants the
-`https://patreon.arkavo.com/attr/tier/value/patron` attribute value.
+Attribute *values* are dynamic (resolved as synthetic values when
+`allow_direct_entitlements` is on), so onboarding a creator requires zero
+policy changes. Declined/former memberships grant nothing, and tiers are
+campaign-qualified so a tier at one creator never satisfies another's gate.
+
+It also surfaces a flattened `.patreon.*` view (`status`, `tier_slug`,
+`campaign_ids`) for coarse legacy subject mappings, derived from the same
+claim.
 
 ## Configuration
 
@@ -36,44 +35,44 @@ Subject mappings in `policy.patreon.yaml` key off these fields - e.g.
 services:
   entityresolution:
     mode: patreon
-    # One of these must be set:
-    creator_access_token: ${PATREON_CREATOR_ACCESS_TOKEN}
-    # ...or for client_credentials refresh:
-    # client_id: ${PATREON_CLIENT_ID}
-    # client_secret: ${PATREON_CLIENT_SECRET}
 
-    # Restrict member lookups to specific campaign ids (leave empty to
-    # auto-list campaigns this token has access to).
-    campaign_ids:
-      - "12345678"
+    # SECURITY: the materialized arkavo_patreon claim is authoritative.
+    # Off by default; enable only when every decision caller reaches the ERS
+    # through a trusted channel (the platform's verified-token path, or a
+    # role:standard PEP that verified the subject token — e.g. the catalog
+    # node). Pin the materializer with trusted_issuer.
+    trust_materialized_claims: true
+    trusted_issuer: https://identity.arkavo.net
 
-    # Return a synthetic free-tier membership when a subject is not a backer
-    # (instead of returning NotFound). Useful so unauthenticated/non-Patreon
-    # traffic still flows through the subject mapping engine.
+    # Namespace for emitted entitlement FQNs (default patreon.arkavo.com).
+    entitlements_namespace: patreon.arkavo.com
+
+    # A subject with no usable Patreon claim resolves as a free follower
+    # instead of NotFound, so non-Patreon traffic still flows through subject
+    # mappings.
     infer_unknown_as_free: true
 
     # JWT claim overrides (defaults shown).
     jwt:
       patreon_user_id_claim: patreon_user_id
-      patreon_access_token_claim: patreon_access_token
       username_claim: preferred_username
       client_id_claim: azp
 ```
 
-## Resolution strategy
+The authorization service must also set `allow_direct_entitlements: true`
+(and typically `enforce_namespaced_entitlements: true`) for the dynamic
+campaign-tier values to be honored.
 
-`ResolveEntities` chooses a lookup path per entity type:
+## Resolution
 
-| Entity type      | Lookup                                                |
-|------------------|-------------------------------------------------------|
-| `UserName`       | Patreon user id (campaign member search)              |
-| `EmailAddress`   | Patreon email (campaign member search)                |
-| `ClientId`       | Treated as Patreon user id                            |
-| `Claims` (JWT)   | In priority: `patreon_access_token` -> `patreon_user_id` -> `email` -> `preferred_username` |
-
-`CreateEntityChainsFromTokens` parses each JWT (signature-unverified, like the
-other ERS modes), emits an `ENVIRONMENT` entity for the `azp` client id, and
-a `SUBJECT` entity whose claims are the `patreon` block above.
+Only a **claims** entity carrying `arkavo_patreon` resolves to a membership.
+Other entity types (username/email/client id) have no Patreon source — live
+lookups were removed — so they resolve as not-found (→ free when
+`infer_unknown_as_free`). `CreateEntityChainsFromTokens` parses each JWT
+(signature verified upstream by the platform authn layer), emits an
+`ENVIRONMENT` entity for the `azp` client id, and a trust-gated `SUBJECT`
+entity carrying the `patreon` block plus the preserved claim for the decision
+flow's second pass.
 
 ## Testing
 
@@ -81,5 +80,4 @@ a `SUBJECT` entity whose claims are the `patreon` block above.
 cd service && go test ./entityresolution/patreon/...
 ```
 
-Tests use an in-memory `MembershipAPI` stub plus an `httptest` server to
-exercise pagination and JSON:API parsing. No outbound Patreon calls are made.
+No outbound Patreon calls — tests exercise the claims-passthrough directly.
